ert-backend/requirements/base.txt
asn1crypto==0.24.0
backports.csv==1.0.7
boto3=1.12.42
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-appconf==1.0.3
django-choices==1.7.0
django-cors-middleware==1.4.0
django-extensions==2.1.9
django-filter==2.1.0
django-fsm==2.6.1
django-fsm-admin==1.2.4
django-fsm-log==1.6.2
django-grappelli==2.10.4
django-import-export==1.2.0
django-money==0.15
django-oauth-toolkit==1.2.0
django-storages==1.9.1
djangorestframework==3.9.4
djangorestframework-gis==0.14
elastic-apm==5.1.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
et-xmlfile==1.0.1
geopy==1.14.0
idna==2.8
jdcal==1.4.1
nonoid==2.0.0
oauthlib==3.0.2
odfpy==1.4.0
openpyxl==2.6.2
pickleshare==0.7.5
Pillow==7.0.0
psycopg2==2.8.2
py-moneyed==0.8.0
pyproj==2.4.2.post1
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
smart-open=1.11.1
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
redis==2.10.6
django-cors-headers==3.1.1
pyfcm==1.4.7
